feat: grab head texture from constants directly
