+++ src/constants/misc.js
@@ -456,98 +456,3 @@ export const DOJO = {
     damage: 0,
   },
 };
-
-export const TROPHY_FISH = {
-  BLOBFISH: {
-    name: "Blobfish",
-    head: "/head/3e77a61f2a25f19bb047be985b965f069d857502881bea3f9682d00bfd5cc3e7",
-    description: "Caught everywhere.",
-  },
-  FLYFISH: {
-    name: "Flyfish",
-    head: "/head/5140c42fc3a1ba2fe77c45b975fa87e8f54a1b1833bc76e6339b4c262304011d",
-    description: "Caught from 8 blocks or higher above lava in the Blazing Volcano.",
-  },
-  GOLDEN_FISH: {
-    name: "Golden Fish",
-    head: "/head/fcfa31d81eae819936834e6664430daf8affbff30b48a1daa7ca1b4c20e3fe7d",
-    description:
-      "Has a chance to spawn after 15 minutes of fishing, increasing linearly until reaching 100% at 20 minutes. The Golden Fish is rolled when your fishing hook is thrown out, regardless of if you catch a fish or not. You are not required to Trophy Fish to catch this one.",
-  },
-  GUSHER: {
-    name: "Gusher",
-    head: "/head/afb8c51bfcd996840010bcce2a3575ae26352e00446d3ec313fcbf1f88108512",
-    description: "Caught within 7-16 minutes after a Blazing Volcano eruption.",
-  },
-  KARATE_FISH: {
-    name: "Karate Fish",
-    head: "/head/901ef47164aba899674cac1e04dda29895ba670807a162e888681c6428e42a83",
-    description:
-      "Caught in the lava pools near the Dojo. Note - Half of the lava pools do not actually count as being in the Dojo area. If you stand in the same place as your bobber and do not see 'Dojo' in the sidebar, you cannot catch the Karate Fish there.",
-  },
-  LAVA_HORSE: {
-    name: "Lava Horse",
-    head: "/head/1176ea86635c4e849469ed694b3c0c3f7ec7677ed0682bee5ef6d59ea669677f",
-    description: "Caught everywhere.",
-  },
-  MANA_RAY: {
-    name: "Mana Ray",
-    head: "/head/fdb2cad06f475b036ad71d61469b8e670e189204350adab6079dc85e1f7d58f2",
-    description: `Caught when you have at least 1,200 ${SYMBOLS.intelligence} Mana.`,
-  },
-  MOLDFIN: {
-    name: "Moldfin",
-    head: "/head/54f33dc405ba447b35926b48d90600302aeebb140ad330d885886cb1029a8af",
-    description: "Caught in the Mystic Marsh.",
-  },
-  OBFUSCATED_FISH_1: {
-    name: "Obfuscated 1",
-    head: "/head/e1f4d91e1bf8d3c4258fe0f28ec2fa40670e25ba06ac4b5cb1abf52a83731a9c",
-    description: "Caught with Corrupted Bait.",
-  },
-  OBFUSCATED_FISH_2: {
-    name: "Obfuscated 2",
-    head: "/head/4fbf7111609f2ec23d9b3f285e1755b62193bd7c3d770576e2b18c48afeb0e29",
-    description: "Caught whilst using Obfuscated 1 as bait.",
-  },
-  OBFUSCATED_FISH_3: {
-    name: "Obfuscated 3",
-    head: "/head/df478663687d16f79b9c32546a7c3ec2736e87ce69779991e52deaf622abd8c2",
-    description: "Caught with Obfuscated 2 as bait.",
-  },
-  SKELETON_FISH: {
-    name: "Skeleton Fish",
-    head: "/head/923e0a25048b60a2cc092f72943159ec170063bb235aa79690ef34ab181d691",
-    description: "Caught in the Burning Desert.",
-  },
-  SLUGFISH: {
-    name: "Slugfish",
-    head: "/head/d82efd885e6e2a964efb857de724b2ef043f1dcbbe618f10ec3742c6f2cecab",
-    description: "Caught when the bobber has been active for at least 30 seconds.",
-  },
-  SOUL_FISH: {
-    name: "Soul Fish",
-    head: "/head/7fe554d346c20c161aa85cfdc1b89779c9f64e726a5bb28ace8078e6594052d7",
-    description: "Caught in the Stronghold.",
-  },
-  STEAMING_HOT_FLOUNDER: {
-    name: "Steaming-Hot Flounder",
-    head: "/head/8b88f88f3053c434660eeb4c7b2344bc21ab52596cea5a66d0f9db8c0e050209",
-    description: "Caught when the bobber is within 2 blocks of a Geyser in the Blazing Volcano.",
-  },
-  SULPHUR_SKITTER: {
-    name: "Sulphur Skitter",
-    head: "/head/4fbf7111609f2ec23d9b3f285e1755b62193bd7c3d770576e2b18c48afeb0e29",
-    description: "Caught when standing within 8 blocks of a Sulphur Ore.",
-  },
-  VANILLE: {
-    name: "Vanille",
-    head: "/head/57120222cce38d53ba69fc6540e97bff9abdbe22ba6068d4ee9af52ecc56842f",
-    description: "Caught when using a Starter Lava Rod with no Enchantments.",
-  },
-  VOLCANIC_STONEFISH: {
-    name: "Volcanic Stonefish",
-    head: "/head/38f89cbaa61ecc99a8321d53f070cef8414efc3eac47bf6fe143056fed9ee8",
-    description: "Caught in the Blazing Volcano.",
-  },
-};
